+++ Server/KFPlugin/KFMatchShard/KFMatchShardModule.cpp
@@ -76,7 +76,7 @@ namespace KFrame
     {
         if ( servername == KFGlobal::Instance()->_app_name && servertype == __KF_STRING__( master ) )
         {
-            std::set< uint64 > objectlist;
+            std::set< uint32 > objectlist;
 
             for ( auto& iter : _kf_match_config->_kf_match_setting._objects )
             {
